Add initial Proxy interface wrapping.
Mark a field of an interface type with @Proxy to wrap it in a proxy implementation.
The field value will be updated to the proxy so that the proxy is passed to external code.
The underlying delegate for the proxy is updated automatically on code updates.
By default, run the interface on the root thread, and wait for the result.
Direct execution and timeout control to follow.